rename Mainactivity to MenuActivity.kt
